+++ screenlockpin.koplugin/plugin/ui/lockscreen/lockscreenframe.lua
@@ -25,9 +25,10 @@ local LockScreenFrame = WidgetContainer:extend {
 }
 
 function LockScreenFrame:init()
+    local uiSettings = pluginSettings.getUiSettings()
     self.lock_widget = ScreenLockWidget:new {
         ui_root = self,
-        scale = 0.5,
+        scale = uiSettings.scale,
         on_update = function(input)
             if input ~= pluginSettings.readPin() then
                 self.lock_widget.state:incFailedCount()